+++ CHANGELOG.md
@@ -5,5 +5,5 @@ All notable changes to this project will be documented in this file, per [the Ke
 ## v1.0
 ### Added
 - Initial plugin release 🎉
-- Sitemaps are generated on the fly
+- Sitemap is generated on the fly
 - Output is saved in an option for fast reading/displaying on the front end